+++ src/test/java/hm1/Fridge.java
@@ -1,27 +1,27 @@
 package hm1;
 
 public class Fridge {
-    private Float height;
-    private Float width;
-    private Float lenght;
+    private Double height;
+    private Double width;
+    private Double lenght;
     private String freezer;
     private Integer numberOfShelves;
-    private Float price;
+    private Double price;
 
 
-    public Float getHeight() {
+    public Double getHeight() {
         return height;
     }
 
-    public void setHeight(Float height) {
+    public void setHeight(Double height) {
         this.height = height;
     }
 
-    public Float getWidth() {
+    public Double getWidth() {
         return width;
     }
 
-    public void setWidth(Float width) {
+    public void setWidth(Double width) {
         this.width = width;
     }
 }
